refactor: Вынесла импорты и файл index.ts в widgets
