--- src/widgets/StatisticSection/ui/CardStatistic/index.ts
@@ -1,7 +1,0 @@
-export { CardStatistic } from './CardStatistic';
-export type {
-  CardStatisticProps,
-  CardType,
-  ChipColor,
-  SuccessRateThresholds,
-} from './CardStatistic.types';
