--- src/widgets/StatisticSection/ui/index.ts
@@ -1,4 +1,0 @@
-export { StatisticSection } from './StatisticSection';
-export type { StatisticSectionProps, StatisticData } from './StatisticSection.types';
-export { CardStatistic } from './CardStatistic';
-export type { CardStatisticProps, CardType } from './CardStatistic';
