+++ src/widgets/index.ts
@@ -1,0 +1,5 @@
+export { StatisticSection } from './StatisticSection/ui/StatisticSection';
+export type {
+  StatisticSectionProps,
+  StatisticData,
+} from './StatisticSection/ui/StatisticSection.types';
